+++ src/Funds/FundDetails/FundType/FundType.js
@@ -1,10 +1,12 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import { KeyValue, NoValue } from '@folio/stripes/components';
+import {
+  KeyValue,
+  NoValue,
+} from '@folio/stripes/components';
 
-const FundType = ({ fundTypeName }) => (
+const FundType = ({ fundTypeName = '' }) => (
   <KeyValue
     label={<FormattedMessage id="ui-finance.fund.information.type" />}
     value={fundTypeName || <NoValue />}
@@ -13,10 +15,6 @@ const FundType = ({ fundTypeName }) => (
 
 FundType.propTypes = {
   fundTypeName: PropTypes.string,
-};
-
-FundType.defaultProps = {
-  fundTypeName: '',
 };
 
 export default FundType;