+++ src/Funds/FundGroups/FieldFundGroups/FieldFundGroups.js
@@ -1,25 +1,28 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { FieldMultiSelectionFinal } from '@folio/stripes-acq-components';
 
-function FieldFundGroups(props) {
+const DEFAULT_DATA_OPTIONS = [];
+
+function FieldFundGroups({
+  dataOptions = DEFAULT_DATA_OPTIONS,
+  name,
+  ...props
+}) {
   return (
     <FieldMultiSelectionFinal
+      dataOptions={dataOptions}
       label={<FormattedMessage id="ui-finance.fund.information.group" />}
+      name={name}
       {...props}
     />
   );
 }
 
 FieldFundGroups.propTypes = {
-  name: PropTypes.string.isRequired,
   dataOptions: PropTypes.arrayOf(PropTypes.any),
-};
-
-FieldFundGroups.defaultProps = {
-  dataOptions: [],
+  name: PropTypes.string.isRequired,
 };
 
 export default FieldFundGroups;